# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API for worm-fertilizer e-commerce + consultations
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  /** e.g. kg, bag */
  unit: string;
  weightKg: number;
  imageUrl: string;
  stock: number;
  active: boolean;
  createdAt: string;
}

export interface ProductInput {
  name: string;
  description: string;
  price: number;
  unit: string;
  weightKg: number;
  imageUrl: string;
  stock: number;
  active?: boolean;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  confirmed: "confirmed",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  customerName: string;
  phone: string;
  address: string;
  city: string;
  /** @nullable */
  notes: string | null;
  productId: number;
  productName: string;
  unitPrice: number;
  quantity: number;
  totalPrice: number;
  status: OrderStatus;
  createdAt: string;
}

export interface OrderInput {
  customerName: string;
  phone: string;
  address: string;
  city: string;
  /** @nullable */
  notes?: string | null;
  productId: number;
  /** @minimum 1 */
  quantity: number;
}

export type OrderStatusUpdateStatus =
  (typeof OrderStatusUpdateStatus)[keyof typeof OrderStatusUpdateStatus];

export const OrderStatusUpdateStatus = {
  pending: "pending",
  confirmed: "confirmed",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export type ConsultationSoilType =
  (typeof ConsultationSoilType)[keyof typeof ConsultationSoilType];

export const ConsultationSoilType = {
  sandy: "sandy",
  clay: "clay",
  silt: "silt",
  loam: "loam",
  rocky: "rocky",
  other: "other",
} as const;

export type ConsultationStatus =
  (typeof ConsultationStatus)[keyof typeof ConsultationStatus];

export const ConsultationStatus = {
  new: "new",
  in_progress: "in_progress",
  answered: "answered",
} as const;

export interface Consultation {
  id: number;
  customerName: string;
  phone: string;
  soilType: ConsultationSoilType;
  crop: string;
  problem: string;
  status: ConsultationStatus;
  createdAt: string;
}

export type ConsultationInputSoilType =
  (typeof ConsultationInputSoilType)[keyof typeof ConsultationInputSoilType];

export const ConsultationInputSoilType = {
  sandy: "sandy",
  clay: "clay",
  silt: "silt",
  loam: "loam",
  rocky: "rocky",
  other: "other",
} as const;

export interface ConsultationInput {
  customerName: string;
  phone: string;
  soilType: ConsultationInputSoilType;
  crop: string;
  problem: string;
}

export type ConsultationStatusUpdateStatus =
  (typeof ConsultationStatusUpdateStatus)[keyof typeof ConsultationStatusUpdateStatus];

export const ConsultationStatusUpdateStatus = {
  new: "new",
  in_progress: "in_progress",
  answered: "answered",
} as const;

export interface ConsultationStatusUpdate {
  status: ConsultationStatusUpdateStatus;
}

export interface AdminLoginInput {
  password: string;
}

export interface AdminLoginResult {
  token: string;
}

export type AdminStatsOrdersByStatusItem = {
  status: string;
  count: number;
};

export interface AdminStats {
  totalOrders: number;
  pendingOrders: number;
  deliveredOrders: number;
  totalRevenue: number;
  totalConsultations: number;
  newConsultations: number;
  totalProducts: number;
  ordersByStatus: AdminStatsOrdersByStatusItem[];
}

export type ActivityItemKind =
  (typeof ActivityItemKind)[keyof typeof ActivityItemKind];

export const ActivityItemKind = {
  order: "order",
  consultation: "consultation",
} as const;

export interface ActivityItem {
  kind: ActivityItemKind;
  id: number;
  title: string;
  subtitle: string;
  createdAt: string;
}

export type AdminTokenParameter = string;