# lib/api-zod/src/generated/types/adminLoginResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API for worm-fertilizer e-commerce + consultations
 * OpenAPI spec version: 0.1.0
 */

export interface AdminLoginResult {
  token: string;
}