# lib/api-zod/src/generated/types/consultation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API for worm-fertilizer e-commerce + consultations
 * OpenAPI spec version: 0.1.0
 */
import type { ConsultationSoilType } from "./consultationSoilType";
import type { ConsultationStatus } from "./consultationStatus";

export interface Consultation {
  id: number;
  customerName: string;
  phone: string;
  soilType: ConsultationSoilType;
  crop: string;
  problem: string;
  status: ConsultationStatus;
  createdAt: Date;
}