# lib/api-zod/src/generated/types/consultationStatusUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API for worm-fertilizer e-commerce + consultations
 * OpenAPI spec version: 0.1.0
 */
import type { ConsultationStatusUpdateStatus } from "./consultationStatusUpdateStatus";

export interface ConsultationStatusUpdate {
  status: ConsultationStatusUpdateStatus;
}