# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API for worm-fertilizer e-commerce + consultations
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  /** e.g. kg, bag */
  unit: string;
  weightKg: number;
  imageUrl: string;
  stock: number;
  active: boolean;
  createdAt: Date;
}